correction

+++ 04models_cc.py
@@ -54,7 +54,7 @@ cuda = "lustre" in current_directory
 if cuda:
     model_name = sys.argv[1]
     part = sys.argv[2]
-    optimizer_lr = int(sys.argv[3]) 
+    optimizer_lr = float(sys.argv[3]) 
     bsize = int(sys.argv[4])  # Convert batch size to an integer
     if torch.cuda.is_available():
         device = torch.device('cuda')  # PyTorch will use the default GPU
